Fixed #10 bufferStringDecorator mutates original data
