+++ packages/nextjs-cache-handler/package.json
@@ -10,7 +10,7 @@
     "type": "git",
     "url": "git+https://github.com/fortedigital/nextjs-cache-handler.git"
   },
-  "version": "1.3.0-canary4",
+  "version": "1.3.0-canary5",
   "type": "module",
   "license": "MIT",
   "description": "Next.js cache handlers",
